+++ include/frontend/analysis/callgraph.hpp
@@ -10,8 +10,9 @@
 
 #pragma once
 
-#include "ast/ast_nodes.hpp"
-#include <graphs/dag.hpp>
+#include "frontend/ast/ast_nodes/function_decl.hpp"
+
+#include "graphs/dag.hpp"
 
 #include <iostream>
 #include <string>
@@ -28,6 +29,7 @@ struct callgraph_value_type final {
 inline bool operator==(const callgraph_value_type &lhs, const callgraph_value_type &rhs) {
   return lhs.m_name == rhs.m_name;
 }
+
 } // namespace paracl::frontend
 
 namespace std {
@@ -39,31 +41,6 @@ template <> struct hash<paracl::frontend::callgraph_value_type> {
 
 namespace paracl::frontend {
 
-class callgraph : public graphs::dag<callgraph_value_type> {
-
-public:
-  callgraph() = default;
-#if 0
-  callgraph(ast::named_function_table *named_table) : m_named_functions{named_table} {
-
-    auto &&ftable = *m_named_functions;
-    call_collector collector;
-    std::vector<ast::i_ast_node *> calls;
-
-    for (auto &&definition : ftable) {
-      calls.clear();
-      auto *def_node = static_cast<ast::function_definition *>(definition.second);
-      auto &&def_name = def_node->name();
-      auto def_name_str = def_name.value();
-      collector.collect(def_node->body(), calls);
-      for (auto *call : calls) {
-        auto *call_node = static_cast<ast::function_call *>(call);
-        auto &&call_name = call_node->name();
-        insert(def_name_str, std::string{call_name});
-      }
-    }
-  }
-#endif
-};
+using callgraph = graphs::dag<callgraph_value_type>;
 
 } // namespace paracl::frontend
